requirements.txt
bcrypt==4.3.0
Flask==3.0.3
Flask-Mail==0.10.0
flasgger==0.9.7.1
flask-cors==5.0.1
google-generativeai==0.8.4
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
pyjwt==2.9.0
pypdf2==3.0.1
python-dotenv==1.0.1
snowflake==1.0.2
snowflake-connector-python==3.12.4
snowflake-ml-python==1.7.2
snowflake-snowpark-python==1.26.0
snowflake._legacy==1.0.0
snowflake.core==1.0.2
snowflake==1.0.2
groq==0.13.1
langchain-mistralai==0.2.4
matplotlib==3.9.4
pandas==2.2.3
pydantic==2.10.4
pypdf==5.1.0
requests==2.31.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-multipart==0.0.20
langsmith==0.3.42